+++ Dal/Categories/Repositories/Interface/ICategoriesRepository.cs
@@ -8,9 +8,9 @@ namespace Dal.Categories.Repositories.Interface;
 public interface ICategoriesRepository : IBaseRepository<CategoriesDal, Guid>
 {
     public Task<List<CategoriesDal>> GetAllUserCategory(string userId, string type);
-    public Task<int?> GetSumCategory(Guid catId, string userId);
-    public Task<int?> GetSumCategoryFromTo(string userId, Guid catId, DateTime from, DateTime to, string type);
+    public Task<decimal?> GetSumCategory(Guid catId, string userId);
+    public Task<decimal?> GetSumCategoryFromTo(string userId, Guid catId, DateTime from, DateTime to, string type);
     public Task<List<OperationDal>> GetOperations(Guid id);
-    public Task<int?> GetSumCurrentMonth(Guid catId, string userId, DateTime date);
+    public Task<decimal?> GetSumCurrentMonth(Guid catId, string userId, DateTime date);
 
 }
